[MAINT]: Burn down RUF067 (non-empty __init__ modules)
Extract logic out of the three package __init__.py files so they
contain only docstrings and re-exports:

- rampart/attacks/__init__.py -> Attacks class moved to
  rampart/attacks/_factory.py.
- rampart/probes/__init__.py -> Probes class moved to
  rampart/probes/_factory.py.
- rampart/payloads/__init__.py -> Payloads class plus the
  _apply_converters_async and _build_text_payload helpers moved
  to rampart/payloads/_facade.py (kept separate from _generator.py
  so the lower-level generator stays focused on LLM I/O).

Public import paths are unchanged (still re-exported via __init__).

- Remove RUF067 from global ruff ignore list.
